service test on more platforms
